core/kotlinx-coroutines-io/src/main/kotlin/kotlinx/coroutines/experimental/io/Closeable.kt
package kotlinx.coroutines.experimental.io

interface Closeable {
    suspend fun aClose()
}

suspend inline fun <T : Closeable?, R> T.use(block: (T) -> R): R {
    var exception: Throwable? = null
    try {
        return block(this)
    } catch (e: Throwable) {
        exception = e
        throw e
    } finally {
        when {
            this == null -> {}
            exception == null -> aClose()
            else ->
                try {
                    aClose()
                } catch (closeException: Throwable) {
                    exception.addSuppressed(closeException)
                }
        }
    }
}
